Move tax-details to billing-tax-details
